# blockchain-api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: KALDRIX Blockchain API
  description: REST and WebSocket API for KALDRIX blockchain operations including DAG management, quantum-resistant cryptography, and PBFT consensus
  version: 1.0.0
  contact:
    name: KALDRIX Team
    email: api@kaldrix.io
    url: https://kaldrix.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.kaldrix.io
    description: Production server
paths:
...
  /api/v1/transactions/submit:
    post:
      summary: Submit Transaction
      description: Submit a new transaction to the blockchain
      tags:
        - Transactions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sender
                - receiver
                - amount
              properties:
                sender:
                  type: string
                  format: address
                  example: kx1q9f5j8g7h6d5s4a3z2x1c9v8b7n6m5k4j3i2h1g
                receiver:
                  type: string
                  format: address
                  example: kx1a2b3c4d5e6f7g8h9i0j1k2l3m4n5o6p7q8r9s0t
                amount:
                  type: number
                  minimum: 0
                  example: 100.5
                data:
                  type: object
                  description: Additional transaction data
                privateKey:
                  type: string
                  description: Private key for signing (optional if using wallet)
      responses:
        '201':
          description: Transaction submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: string
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  message:
                    type: string
        '400':
          description: Invalid transaction data
  /api/v1/transactions/{transactionId}:
    get:
      summary: Get Transaction
      description: Retrieve a specific transaction by ID
      tags:
        - Transactions
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
  /api/v1/transactions/:
    get:
      summary: Get Transactions
      description: Retrieve transactions with optional filtering
      tags:
        - Transactions
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sender
          in: query
          schema:
            type: string
            format: address
        - name: receiver
          in: query
          schema:
            type: string
            format: address
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, failed]
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  filters:
                    type: object

  # Wallet Operations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Transaction:
      type: object
      required:
        - id
        - sender
        - receiver
        - amount
        - timestamp
        - signature
      properties:
        id:
          type: string
          example: "tx_0x123abc456def789ghi012jkl345mno678pqr901stu"
        sender:
          type: string
          format: address
          example: "kx1q9f5j8g7h6d5s4a3z2x1c9v8b7n6m5k4j3i2h1g"
        receiver:
          type: string
          format: address
          example: "kx1a2b3c4d5e6f7g8h9i0j1k2l3m4n5o6p7q8r9s0t"
        amount:
          type: number
          example: 100.5
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        signature:
          type: string
          example: "0x123abc456def789ghi012jkl345mno678pqr901stu..."
        parents:
          type: array
          items:
            type: string
          description: Parent transaction IDs in DAG
        status:
          type: string
          enum: [pending, confirmed, failed]
          example: "confirmed"
        fee:
          type: number
          example: 0.001
        data:
          type: object
          description: Additional transaction data
...
security:
  - bearerAuth: []
  - apiKeyAuth: []